Make the package a single source and fix the entry points
The branch had every module twice: the real code at the top level and a
loader in cif2qewan/ that read it back out of the parent directory. Only
cif2qewan was listed in py_modules, so after "pip install" the band_comp and
wannier_conv console scripts raised FileNotFoundError, and band_comp.main()
took an argument the console script never passed.

- Move the three modules into the package; drop the loaders.
- Split band_comp.main() into a zero-argument entry point and
  plot_band_comparison(output_dir).
- Keep one copy of the pseudopotential tables, inside the package, so they
  ship with the wheel.
- Drop setup.py; pyproject.toml is now the only metadata, so the two cannot
  disagree about python_requires any more.
- Declare GPLv2, matching LICENSE, and version 0.1.0 rather than 1.0.0.
- Drop the cif2qewan@example.com placeholder address.

Verified with a built wheel in a clean virtualenv: all three console scripts
run and the CSV tables resolve from site-packages.
