+++ MANIFEST.in
@@ -1,23 +1,11 @@
-# Include package data files
 include README.md
 include LICENSE
-include requirements.txt
-include pyproject.toml
-include setup.py
-include cif2qewan.py
+include CHANGELOG.md
 include cif2qewan.toml
-include *.csv
-include *.sh
+include submit_all.sh
 
-# Include all Python files
-recursive-include cif2qewan *.py
+recursive-include cif2qewan *.py *.csv
 
-# Include configuration files
-recursive-include cif2qewan *.toml
-recursive-include cif2qewan *.csv
-recursive-include cif2qewan *.sh
-
-# Exclude test files from distribution
 recursive-exclude tests *
 recursive-exclude * __pycache__
 recursive-exclude * *.py[co]
