--- band_comp.py
@@ -1,302 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Band structure comparison tool for Quantum ESPRESSO and Wannier90.
-
-This module provides functionality to compare band structures calculated using
-Quantum ESPRESSO (PWSCF) with those obtained from Wannier90 interpolation.
-The comparison is performed by plotting both band structures on the same graph
-with proper k-path normalization and Fermi energy alignment.
-
-Author: Generated by cif2qewan
-"""
-
-# Standard library imports
-import argparse
-import os.path
-
-# Third-party imports
-import matplotlib
-import numpy as np
-
-# Configure matplotlib for non-interactive backend.
-# This must happen before pyplot is imported so that the script also works
-# on machines without a display.
-matplotlib.use("Agg")
-
-from matplotlib import pyplot as plt  # noqa: E402
-
-def get_ef_from_scfout(scfout: str) -> float:
-    """
-    Extract Fermi energy from Quantum ESPRESSO scf.out file.
-
-    Parameters
-    ----------
-    scfout : str
-        Path to the Quantum ESPRESSO scf.out file.
-
-    Returns
-    -------
-    float
-        Fermi energy in eV.
-
-    Raises
-    ------
-    AssertionError
-        If the scf.out file does not exist.
-
-    Examples
-    --------
-    >>> ef = get_ef_from_scfout("scf.out")
-    >>> print(f"Fermi energy: {ef} eV")
-    """
-    assert os.path.exists(scfout), f"File {scfout} does not exist"
-
-    ef = 0.0
-    with open(scfout, "r") as fp:
-        for line in fp.readlines():
-            # Look for Fermi energy line in QE output.
-            # Do not break: scf.out can contain several "Fermi" lines and the
-            # converged value is the last one.
-            if "Fermi" in line:
-                # Extract Fermi energy from the second-to-last column
-                ef = float(line.split()[-2])
-    
-    return ef
-
-
-def get_band_data(band_file: str) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Extract band structure data from Quantum ESPRESSO band.gnu file.
-
-    The function reads band structure data and reshapes it into a proper
-    format for plotting. The k-path coordinates are normalized to [0,1].
-
-    Parameters
-    ----------
-    band_file : str
-        Path to the band.gnu file generated by Quantum ESPRESSO.
-
-    Returns
-    -------
-    tuple[np.ndarray, np.ndarray]
-        A tuple containing:
-        - x: Normalized k-path coordinates (shape: [nk])
-        - y: Band energies (shape: [nk, nband])
-
-    Examples
-    --------
-    >>> x, y = get_band_data("bands.out.gnu")
-    >>> print(f"Number of k-points: {len(x)}")
-    >>> print(f"Number of bands: {y.shape[1]}")
-    """
-    # Load band structure data from file
-    data = np.loadtxt(band_file)
-
-    # Extract k-path coordinates (first column)
-    x = data[:, 0]
-
-    # Count number of bands by counting k=0 points (band separators)
-    nband = np.sum(x == 0)
-
-    # Calculate number of k-points per band
-    nk = int(len(x) / nband)
-
-    # Normalize k-path to [0,1] range
-    x = x[0:nk] / max(x)
-
-    # Reshape energies and transpose to get [nk, nband] format
-    y = np.transpose(data[:, 1].reshape(nband, nk))
-
-    return x, y
-
-
-def get_froz_max(pwscf_win: str) -> float:
-    """
-    Extract dis_froz_max parameter from Wannier90 input file.
-
-    Parameters
-    ----------
-    pwscf_win : str
-        Path to the Wannier90 input file (pwscf.win).
-
-    Returns
-    -------
-    float
-        The dis_froz_max parameter value, default -200.0 if not found.
-
-    Examples
-    --------
-    >>> froz_max = get_froz_max("pwscf.win")
-    >>> print(f"dis_froz_max: {froz_max}")
-    """
-    dis_froz_max = -200.0  # Default value
-
-    with open(pwscf_win, "r") as fp:
-        for line in fp.readlines():
-            # Look for dis_froz_max parameter in Wannier90 input.
-            # Do not break: the last assignment in the file is the effective one.
-            if "dis_froz_max" in line:
-                dis_froz_max = float((line.split())[-1])
-    
-    return dis_froz_max
-
-
-def get_klabel(wannier_band_gnu: str) -> tuple[np.ndarray, list[str]]:
-    """
-    Extract k-point labels and positions from Wannier90 band.gnu file.
-
-    This function parses the gnuplot xtics information from the Wannier90
-    band structure file to extract k-point labels and their positions.
-
-    Parameters
-    ----------
-    wannier_band_gnu : str
-        Path to the Wannier90 band.gnu file.
-
-    Returns
-    -------
-    tuple[np.ndarray, list[str]]
-        A tuple containing:
-        - x_list: Normalized k-point positions (shape: [num_klabel])
-        - label_list: K-point labels with LaTeX formatting
-
-    Examples
-    --------
-    >>> x_pos, labels = get_klabel("pwscf_band.gnu")
-    >>> print(f"K-point labels: {labels}")
-    """
-    xtics = ""
-
-    # Read the gnuplot file to find xtics line
-    with open(wannier_band_gnu, "r") as fp:
-        for line in fp.readlines():
-            if "set xtics" in line:
-                # Extract xtics information
-                xtics = line.replace("set xtics", "").strip()
-                break
-
-    # Parse xtics string to extract labels and positions
-    xtics = xtics[1:-1].split(",")  # Remove parentheses and split by comma
-    x_list = []
-    label_list = []
-
-    for xtic in xtics:
-        # Split each xtic into label and position
-        label, pos = xtic.split()
-        x_list.append(float(pos))
-
-        # Format labels with LaTeX symbols for Greek letters
-        formatted_label = (
-            label[1:-1]  # Remove quotes
-            .replace("G", "$\\Gamma$")
-            .replace("S", "$\\Sigma$")
-        )
-        label_list.append(formatted_label)
-
-    # Normalize k-point positions to [0,1] range
-    x_list = np.array(x_list) / max(x_list)
-
-    return x_list, label_list
-
-
-def main(output_dir: str) -> None:
-    """
-    Generate band structure comparison plot between QE and Wannier90.
-
-    This function creates a comparison plot showing both Quantum ESPRESSO
-    and Wannier90 band structures on the same graph. The plot includes
-    proper k-point labels, Fermi energy alignment, and vertical lines
-    marking high-symmetry points.
-
-    Parameters
-    ----------
-    output_dir : str
-        Directory path where the output plots will be saved.
-
-    Examples
-    --------
-    >>> main("./output")
-    """
-    # Define input file paths
-    scfout = "scf.out"  # QE SCF output file
-    wannier_band = "pwscf_band.dat"  # Wannier90 band data
-    wannier_band_gnu = "pwscf_band.gnu"  # Wannier90 gnuplot file
-    pwscf_band = "../band/bands.out.gnu"  # QE band structure data
-
-    # Load band structure data from both sources
-    x, y = get_band_data(wannier_band)  # Wannier90 data
-    x_qe, y_qe = get_band_data(pwscf_band)  # QE data
-
-    # Get Fermi energy and k-point labels
-    ef = get_ef_from_scfout(scfout)
-    klabel = get_klabel(wannier_band_gnu)
-
-    # Configure matplotlib for publication-quality plots
-    plt.rcParams["font.size"] = 16
-
-    # Create the comparison plot
-    plt.title("Red=QE, Black=Wannier")
-
-    # Plot QE bands in red
-    plt.plot(x_qe, y_qe - ef, c="r", lw=1, label="QE")
-
-    # Plot Wannier90 bands in black
-    plt.plot(x, y - ef, c="k", lw=0.5, label="Wannier")
-
-    # Set axis labels and limits
-    plt.ylabel(r"$E - E_{\mathrm{F}}$[eV]")
-    plt.xticks(klabel[0], klabel[1])  # Set k-point labels
-    plt.xlim([0, 1])  # Normalized k-path range
-
-    # Calculate y-axis limits with 5% padding
-    y_min = np.min(y - ef)
-    y_max = np.max(y - ef)
-    py_min = y_min - 0.05 * (y_max - y_min)
-    py_max = y_max + 0.05 * (y_max - y_min)
-    plt.ylim([py_min, py_max])
-
-    # Add vertical lines at high-symmetry points
-    plt.vlines(klabel[0], py_min, py_max, colors="black", linewidth=1.0, zorder=3)
-
-    # Save plots in multiple formats
-    plt.savefig(f"{output_dir}/band_compare.png", bbox_inches="tight")
-    plt.savefig(f"{output_dir}/band_compare.eps", bbox_inches="tight")
-
-
-if __name__ == "__main__":
-    """
-    Command-line interface for band structure comparison.
-
-    This script can be run from the command line to generate band structure
-    comparison plots between Quantum ESPRESSO and Wannier90 calculations.
-
-    Usage:
-        python band_comp.py -o <output_directory>
-
-    Arguments:
-        -o, --odir: Output directory for the comparison plots
-    """
-    # Set up command-line argument parser
-    parser = argparse.ArgumentParser(
-        description="Compare band structures from QE and Wannier90",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  python band_comp.py -o ./plots
-  python band_comp.py -o /path/to/output
-        """,
-    )
-    parser.add_argument(
-        "-o",
-        dest="odir",
-        required=True,
-        help="Output directory for the comparison plots",
-    )
-
-    # Parse command-line arguments
-    args = parser.parse_args()
-    output_dir = args.odir
-
-    # Generate the band structure comparison
-    main(output_dir)
