--- cif2qewan.py
@@ -1,1051 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-CIF to Quantum ESPRESSO Wannier90 input generator.
-
-This module provides functionality to generate Quantum ESPRESSO and Wannier90
-input files from CIF (Crystallographic Information File) structures. It supports
-both standard DFT calculations and advanced features like spin-orbit coupling
-and magnetic calculations.
-
-Usage:
-    cif2qewan.py <cif_file> <toml_file> [--so] [--mag]
-
-Options:
-    --so     input including spin-orbit couplings
-    --mag    ferromagnetic calculations
-
-Author: Generated by cif2qewan
-"""
-
-# Standard library imports
-import itertools
-import os
-import re
-import warnings
-
-# Third-party imports
-import numpy as np
-import pandas as pd
-import toml
-from docopt import docopt
-from pymatgen.core.periodic_table import Element
-
-
-class qe_wannier_in:
-    """
-    Generate Quantum ESPRESSO and Wannier90 input files from CIF structures.
-
-    This class handles the conversion of CIF files to Quantum ESPRESSO input
-    files and generates corresponding Wannier90 input files for band structure
-    interpolation. It supports various calculation types including spin-orbit
-    coupling and magnetic calculations.
-
-    Parameters
-    ----------
-    cif_file : str
-        Path to the input CIF file.
-    toml_file : str
-        Path to the configuration TOML file.
-    so : bool
-        Whether to include spin-orbit coupling.
-    mag : bool
-        Whether to perform magnetic calculations.
-
-    Attributes
-    ----------
-    cif2cell_path : str
-        Path to the cif2cell executable.
-    scf_k_resolution : float
-        K-point resolution for SCF calculations.
-    degauss : float
-        Gaussian smearing parameter.
-    pseudo_dir : str
-        Directory containing pseudopotential files.
-    pp_list_path : str
-        Path to pseudopotential list file.
-    info_pw2wan : dict
-        Configuration for pw2wannier90 interface.
-    so : bool
-        Spin-orbit coupling flag.
-    mag : bool
-        Magnetic calculation flag.
-    lines : list[str]
-        Lines from cif2cell output.
-    system_str : str
-        Quantum ESPRESSO system namelist.
-    control_str : str
-        Quantum ESPRESSO control namelist.
-    electrons_str : str
-        Quantum ESPRESSO electrons namelist.
-    """
-
-    def __init__(self, cif_file: str, toml_file: str, so: bool, mag: bool) -> None:
-        """
-        Initialize the QE Wannier input generator.
-
-        Parameters
-        ----------
-        cif_file : str
-            Path to the input CIF file.
-        toml_file : str
-            Path to the configuration TOML file.
-        so : bool
-            Whether to include spin-orbit coupling.
-        mag : bool
-            Whether to perform magnetic calculations.
-        """
-        # Load configuration from TOML file
-        info = toml.load(toml_file)
-        self.cif2cell_path = info["cif2cell_path"]
-        self.scf_k_resolution = info["scf_k_resolution"]
-        self.degauss = info["degauss"]
-        self.pseudo_dir = info["pseudo_dir"]
-        self.pp_list_path = info["pp_list_path"]
-        self.info_pw2wan = info["pw2wan"]
-
-        # Store calculation flags
-        self.so = so
-        self.mag = mag
-
-        # Generate initial QE input using cif2cell
-        self.lines = self.cif2cell_scf_in(cif_file)
-
-        # Initialize QE namelist strings
-        self.system_str = "&system\n"
-        self.control_str = "&control\n"
-        self.electrons_str = "&electrons\n"
-
-        # Parse system information and set up calculations
-        ntyp, nat = self.read_set_system()
-        ecut_wfc, ecut_rho = self.read_set_pseudo_other(ntyp, nat)
-        self.set_system2(ntyp, ecut_wfc, ecut_rho)
-        self.set_control()
-        self.set_electrons(conv_thr="1.0d-8")
-
-    def cif2cell_scf_in(self, cif_file: str) -> list[str]:
-        """
-        Generate Quantum ESPRESSO input file using cif2cell.
-
-        This method calls the cif2cell utility to convert a CIF file to a
-        Quantum ESPRESSO input file with appropriate k-point sampling.
-
-        Parameters
-        ----------
-        cif_file : str
-            Path to the input CIF file.
-
-        Returns
-        -------
-        list[str]
-            Lines from the generated Quantum ESPRESSO input file.
-
-        Examples
-        --------
-        >>> lines = self.cif2cell_scf_in("structure.cif")
-        >>> print(f"Generated {len(lines)} lines")
-        """
-        cif_scf_in = "cif_scf.in"
-
-        # Only generate if file doesn't exist
-        if not os.path.exists(cif_scf_in):
-            # Escape parentheses in filename for shell command
-            escaped_cif_file = cif_file.replace("(", "\\(").replace(")", "\\)")
-
-            # Call cif2cell with appropriate parameters
-            cmd = (
-                f"{self.cif2cell_path} -p pwscf --setup-all "
-                f"--k-resolution={self.scf_k_resolution:.3f} "
-                f"--print-digits=10 -o {cif_scf_in} {escaped_cif_file}"
-            )
-            os.system(cmd)
-
-        return open(cif_scf_in).readlines()
-
-    def read_set_system(self) -> tuple[int, int]:
-        """
-        Parse system parameters from cif2cell output.
-
-        This method extracts key system parameters like lattice type (ibrav),
-        lattice constant (A), number of atom types (ntyp), and number of atoms (nat)
-        from the cif2cell-generated input file.
-
-        Returns
-        -------
-        tuple[int, int]
-            A tuple containing (ntyp, nat) where:
-            - ntyp: Number of atom types
-            - nat: Number of atoms
-
-        Examples
-        --------
-        >>> ntyp, nat = self.read_set_system()
-        >>> print(f"System has {ntyp} atom types and {nat} atoms")
-        """
-        ntyp = 0
-        nat = 0
-
-        for line in self.lines:
-            # Extract lattice type (ibrav)
-            if "ibrav" in line:
-                self.system_str += line
-
-            # Extract lattice constant and store for later use
-            if "A =" in line:
-                self.system_str += line
-                self.alat = float(line.split("=")[1])
-
-            # Extract number of atom types
-            if "ntyp =" in line:
-                self.system_str += line
-                ntyp = int(line.split("=")[1])
-
-            # Extract number of atoms
-            if "nat =" in line:
-                self.system_str += line
-                nat = int(line.split("=")[1])
-
-        # Fail loudly rather than continuing with an empty system: a missing
-        # ntyp/nat means cif2cell did not produce a usable input file.
-        if ntyp == 0 or nat == 0:
-            raise ValueError(
-                "could not read ntyp/nat from cif2cell output "
-                "(got ntyp={}, nat={})".format(ntyp, nat)
-            )
-
-        return ntyp, nat
-
-    def set_system2(self, ntyp: int, ecut_wfc: float, ecut_rho: float) -> None:
-        """
-        Set additional system parameters for Quantum ESPRESSO.
-
-        This method adds energy cutoffs, smearing parameters, and spin-related
-        settings to the system namelist based on the calculation type.
-
-        Parameters
-        ----------
-        ntyp : int
-            Number of atom types in the system.
-        ecut_wfc : float
-            Wavefunction cutoff energy in Ry.
-        ecut_rho : float
-            Charge density cutoff energy in Ry.
-
-        Examples
-        --------
-        >>> self.set_system2(2, 40.0, 200.0)
-        """
-        # Set energy cutoffs
-        self.system_str += f"  ecutwfc = {ecut_wfc}\n"
-        self.system_str += f"  ecutrho = {ecut_rho}\n"
-
-        # Set smearing parameters for metallic systems
-        self.system_str += "  occupations = 'smearing'\n"
-        self.system_str += "  smearing = 'm-p'\n"
-        self.system_str += f"  degauss = {self.degauss:.3f}\n"
-
-        # Handle magnetic calculations
-        if self.mag:
-            self.system_str += "  nspin = 2\n"
-            # Set initial magnetization for each atom type
-            for i in range(ntyp):
-                self.system_str += f"  starting_magnetization({i+1}) = 3.0\n"
-
-        # Handle spin-orbit coupling
-        elif self.so:
-            self.system_str += "  lspinorb = .true.\n"
-            self.system_str += "  noncolin = .true.\n"
-
-    def set_control(self) -> None:
-        """
-        Set control parameters for Quantum ESPRESSO.
-
-        This method configures the control namelist with standard settings
-        for SCF calculations, including file paths and output options.
-
-        Examples
-        --------
-        >>> self.set_control()
-        """
-        self.control_str += "  calculation = 'scf'\n"
-        self.control_str += "  restart_mode = 'from_scratch'\n"
-        self.control_str += "  prefix = 'pwscf'\n"
-        self.control_str += "  tstress = .true.\n"
-        self.control_str += "  tprnfor = .true.\n"
-        self.control_str += f"  pseudo_dir = '{self.pseudo_dir}'\n"
-        self.control_str += "  outdir = './work'\n"
-        self.control_str += "  wf_collect = .true.\n"
-        self.control_str += "  disk_io = 'low'\n"
-
-    def set_electrons(self, conv_thr: str) -> None:
-        """
-        Set electron-related parameters for Quantum ESPRESSO.
-
-        This method configures the electrons namelist with convergence
-        and mixing parameters for SCF calculations.
-
-        Parameters
-        ----------
-        conv_thr : str
-            Convergence threshold for SCF calculations.
-
-        Examples
-        --------
-        >>> self.set_electrons("1.0d-8")
-        """
-        self.electrons_str += "  mixing_mode = 'plain'\n"
-        self.electrons_str += "  mixing_beta = 0.1\n"
-        self.electrons_str += f"  conv_thr = {conv_thr}\n"
-
-    def read_set_pseudo_other(self, ntyp: int, nat: int) -> tuple[float, float]:
-        """
-        Parse pseudopotential information and set up atomic species.
-
-        This method reads the atomic species section from the cif2cell output,
-        assigns appropriate pseudopotentials, and determines energy cutoffs
-        based on the pseudopotential requirements.
-
-        Parameters
-        ----------
-        ntyp : int
-            Number of atom types in the system.
-        nat : int
-            Number of atoms in the system.
-
-        Returns
-        -------
-        tuple[float, float]
-            A tuple containing (ecut_wfc, ecut_rho) where:
-            - ecut_wfc: Maximum wavefunction cutoff energy in Ry
-            - ecut_rho: Maximum charge density cutoff energy in Ry
-
-        Examples
-        --------
-        >>> ecut_wfc, ecut_rho = self.read_set_pseudo_other(2, 8)
-        >>> print(f"Energy cutoffs: {ecut_wfc} Ry, {ecut_rho} Ry")
-        """
-        ecut_wfc = 0.0
-        ecut_rho = 0.0
-
-        # Initialize pseudopotential list handler
-        pslist = pseudo_list(self.pseudo_dir, self.pp_list_path)
-
-        # Initialize strings for QE input
-        self.pseudo_str = "ATOMIC_SPECIES\n"
-        self.projection_str = ""
-
-        # Initialize dictionaries for Wannier90 setup
-        num_wann_dict = {}
-        nexclude_dict = {}
-
-        # Initialize atom information lists
-        self.atom_list = []
-        self.atom_pos_list = []
-        self.num_wann = 0
-        self.nexclude = 0
-        # Parse the cif2cell output line by line
-        for i, line in enumerate(self.lines):
-            # Extract cell parameters (lattice vectors)
-            if "CELL_PARAM" in line:
-                self.cellparam_str = "".join(self.lines[i : i + 4])
-                # Store lattice vectors as numpy arrays
-                self.a1 = np.array([float(x) for x in self.lines[i + 1].split()])
-                self.a2 = np.array([float(x) for x in self.lines[i + 2].split()])
-                self.a3 = np.array([float(x) for x in self.lines[i + 3].split()])
-
-            # Extract atomic positions
-            if "ATOMIC_POSITIONS" in line:
-                self.atompos_str = "".join(self.lines[i : i + nat + 1])
-                self.wan_atompos_str = "".join(self.lines[i + 1 : i + nat + 1])
-
-                # Parse each atom's position and type
-                for j in range(nat):
-                    atm = self.lines[i + j + 1].split()[0]
-                    self.atom_list.append(atm)
-                    self.atom_pos_list.append(
-                        [float(x) for x in (self.lines[i + j + 1].split()[1:4])]
-                    )
-                    # Accumulate Wannier90 parameters
-                    self.num_wann += num_wann_dict[atm]
-                    self.nexclude += nexclude_dict[atm]
-
-            # Extract k-point mesh
-            if "K_POINTS" in line:
-                self.kpoints_str = "".join(self.lines[i : i + 2])
-                self.kmesh = [int(x) for x in self.lines[i + 1].split()[0:3]]
-
-            # Process atomic species and assign pseudopotentials
-            if "ATOMIC_SPECIES" in line:
-                for j in range(ntyp):
-                    line = self.lines[i + j + 1]
-                    atm = line[:5].strip()  # Extract atom type
-                    ps = pslist.pseudo(atm)  # Get pseudopotential info
-
-                    # Update energy cutoffs based on pseudopotential requirements
-                    ecut_wfc = max(ecut_wfc, ps[4])
-                    ecut_rho = max(ecut_rho, ps[5])
-
-                    # Check if ecut_rho is appropriate
-                    if ecut_rho < 4 * ecut_wfc:
-                        warnings.warn(
-                            f"ecut_rho should be bigger than 4*ecut_wfc, "
-                            f"but {ecut_rho} < {4*ecut_wfc}."
-                        )
-
-                    # Add pseudopotential to ATOMIC_SPECIES section
-                    self.pseudo_str += re.sub("[A-Za-z]+_PSEUDO", ps[0], line)
-
-                    # Handle spin-orbit coupling pseudopotentials
-                    if self.so and not self.mag:
-                        self.pseudo_str = self.pseudo_str.replace(".pbe", ".rel-pbe")
-                        self.pseudo_str = self.pseudo_str.replace("_sr", "_fr")
-
-                    # Set up Wannier90 projections
-                    if ps[2] != "":
-                        self.projection_str += f"{atm}:{','.join(list(ps[2]))}\n"
-
-                    # Store Wannier90 parameters
-                    num_wann_dict[atm] = ps[3]
-                    nexclude_dict[atm] = ps[1]
-
-        return ecut_wfc, ecut_rho
-
-    def convert2nscf(self) -> None:
-        """
-        Convert SCF input to NSCF input for Wannier90 calculations.
-
-        This method modifies the Quantum ESPRESSO input to perform a non-self-consistent
-        field (NSCF) calculation with a denser k-point mesh suitable for Wannier90
-        interpolation. It also sets up the appropriate number of bands and k-points.
-
-        Examples
-        --------
-        >>> self.convert2nscf()
-        """
-        # Change calculation type from SCF to NSCF
-        self.control_str = self.control_str.replace("'scf'", "'nscf'")
-
-        # Add NSCF-specific system parameters
-        system_add_str = "  nosym = .true.\n"
-
-        # Set number of bands based on calculation type
-        if self.so or self.mag:
-            # For spin-orbit or magnetic calculations, double the bands
-            system_add_str += f"  nbnd = {(self.nexclude + self.num_wann*3)*2}\n"
-        else:
-            # Standard calculation
-            system_add_str += f"  nbnd = {self.nexclude + self.num_wann*3}\n"
-
-        self.system_str = self.system_str.replace(
-            "&system\n", "&system\n" + system_add_str
-        )
-
-        # Handle magnetic calculations
-        if self.mag:
-            if self.so:
-                mag_str = "  lspinorb = .true.\n"
-            else:
-                mag_str = "  lspinorb = .false.\n"
-            mag_str += "  noncolin = .true.\n"
-            mag_str += "  lforcet = .true.\n"
-            mag_str += "  angle1 = 0\n"
-            mag_str += "  angle2 = 0\n"
-            self.system_str = self.system_str.replace("  nspin = 2\n", mag_str)
-
-            # Update pseudopotentials for spin-orbit coupling
-            if self.so:
-                self.pseudo_str = self.pseudo_str.replace(".pbe", ".rel-pbe")
-                self.pseudo_str = self.pseudo_str.replace("_sr", "_fr")
-
-        # Set stricter convergence for NSCF
-        self.electrons_str = re.sub(
-            "  conv_thr.*\n", "  conv_thr = 1.d-10\n", self.electrons_str
-        )
-
-        # Generate denser k-point mesh for NSCF
-        self.nscfk = [min(max(nk, 4), 8) for nk in self.kmesh]
-        self.kpoints_str = "K_POINTS {crystal}\n"
-        self.kpoints_str += f"{np.prod(self.nscfk)}\n"
-        self.wan_kmesh = ""
-
-        # Generate k-points for the mesh
-        for kx, ky, kz in itertools.product(
-            range(self.nscfk[0]), range(self.nscfk[1]), range(self.nscfk[2])
-        ):
-            # Add k-point with weight
-            self.kpoints_str += (
-                f"{kx/self.nscfk[0]:15.10f} "
-                f"{ky/self.nscfk[1]:15.10f} "
-                f"{kz/self.nscfk[2]:15.10f} "
-                f"{1.0/np.prod(self.nscfk):15.10f}\n"
-            )
-
-            # Store k-point for Wannier90
-            self.wan_kmesh += (
-                f"{kx/self.nscfk[0]:15.10f} "
-                f"{ky/self.nscfk[1]:15.10f} "
-                f"{kz/self.nscfk[2]:15.10f}\n"
-            )
-
-    def shift_k_nscf(self) -> None:
-        """
-        Create NSCF input with shifted k-points for Wannier90 convergence check.
-
-        This method generates an NSCF calculation with k-points shifted by half
-        a mesh spacing in each direction. This is used to check the convergence
-        of Wannier90 calculations by comparing results at different k-points.
-
-        Note
-        ----
-        This method must be called after convert2nscf().
-
-        Examples
-        --------
-        >>> self.convert2nscf()
-        >>> self.shift_k_nscf()
-        """
-        # Increase verbosity for detailed output
-        self.control_str += "  verbosity = 'high'\n"
-
-        # Remove nosym for shifted k-points
-        self.system_str = self.system_str.replace("  nosym = .true.\n", "")
-
-        # Adjust number of bands for shifted calculation
-        if self.so or self.mag:
-            nbnd = (self.nexclude + int(self.num_wann * 1.5)) * 2
-        else:
-            nbnd = self.nexclude + int(self.num_wann * 1.5)
-        self.system_str = re.sub("  nbnd.*\n", f"  nbnd = {nbnd}\n", self.system_str)
-
-        # Remove full accuracy diagonalization and relax convergence
-        self.electrons_str = self.electrons_str.replace(
-            "  diago_full_acc = .true.\n", ""
-        )
-        self.electrons_str = re.sub(
-            "  conv_thr.*\n", "  conv_thr = 1.d-8\n", self.electrons_str
-        )
-
-        # Use automatic k-point generation with shift
-        self.kpoints_str = "K_POINTS {automatic}\n"
-        self.kpoints_str += f"{self.nscfk[0]} {self.nscfk[1]} {self.nscfk[2]}  1 1 1\n"
-
-    def convert2band(self) -> None:
-        """
-        Convert NSCF input to band structure calculation input.
-
-        This method modifies the input to perform a band structure calculation
-        along high-symmetry lines in the Brillouin zone. It uses the k-path
-        generated by seekpath to create a proper band structure plot.
-
-        Examples
-        --------
-        >>> self.calc_bands_seekpath()
-        >>> self.convert2band()
-        """
-        # Change calculation type to bands
-        self.control_str = self.control_str.replace("'nscf'", "'bands'")
-
-        # Set up k-points for band structure calculation
-        self.kpoints_str = "K_POINTS {crystal_b}\n"
-        # Count non-empty tick labels
-        nkpoints = len(self.tick_labels) - self.tick_labels.count("")
-        self.kpoints_str += f"{nkpoints}\n"
-
-        # Add k-points along the band path
-        j = 0
-        for i in range(len(self.tick_labels)):
-            if self.tick_labels[i] != "":
-                # Format k-point with label and number of points
-                kstr = (
-                    f"{self.tick_locs[i][0]:15.10f} "
-                    f"{self.tick_locs[i][1]:15.10f} "
-                    f"{self.tick_locs[i][2]:15.10f}  "
-                    f"{self.band_nks[j]:>4d}    !  "
-                    f"{self.tick_labels[i]}\n"
-                )
-                self.kpoints_str += kstr
-                j += 1
-
-    def calc_bands_seekpath(self) -> None:
-        """
-        Calculate band structure k-path using seekpath library.
-
-        This method uses the seekpath library to generate an optimal k-path
-        for band structure calculations. If seekpath is not available or
-        incompatible, it falls back to a simple cubic k-path.
-
-        Examples
-        --------
-        >>> self.calc_bands_seekpath()
-        >>> print(f"K-path has {len(self.tick_labels)} points")
-        """
-        try:
-            import seekpath
-
-            if seekpath.__version__ < "2.1.0":
-                raise ImportError
-
-        except ImportError:
-            print("Failed to import seekpath (>= 2.1.0). Simple kpath is used instead.")
-            # Fallback to simple cubic k-path
-            self.tick_labels = ["R", "G", "X", "M", "G"]
-            self.tick_locs = [
-                [0.5, 0.5, 0.5],
-                [0.0, 0.0, 0.0],
-                [0.5, 0.0, 0.0],
-                [0.5, 0.5, 0.0],
-                [0.0, 0.0, 0.0],
-            ]
-            self.band_nks = np.full(len(self.tick_labels), 20)
-            return
-
-        # Prepare input for seekpath
-        cell = np.array([self.a1, self.a2, self.a3])
-        pos = self.atom_pos_list
-        z = [Element(s).Z for s in self.atom_list]
-
-        # Get k-path from seekpath
-        kpath = seekpath.getpaths.get_explicit_k_path_orig_cell([cell, pos, z])
-
-        # Extract k-points and labels
-        self.kpoints_rel = kpath["explicit_kpoints_rel"]
-        kpoints_labels = kpath["explicit_kpoints_labels"]
-
-        # Process k-points and labels
-        hsp_indices = []
-        discon_hsps = []
-        self.tick_locs = []
-        self.tick_labels = []
-
-        for i, label in enumerate(kpoints_labels):
-            if label == "":
-                continue
-
-            # Standardize label names
-            label = label.replace("GAMMA", "G")
-            label = label.replace("SIGMA", "S")
-
-            # Handle discontinuous paths
-            if i != 0 and kpoints_labels[i - 1] != "":
-                self.tick_labels.extend(["", label])
-                self.tick_locs.extend([np.array([0.0, 0.0, 0.0]), self.kpoints_rel[i]])
-                discon_hsps.append(i)
-            else:
-                self.tick_labels.append(label)
-                self.tick_locs.append(self.kpoints_rel[i])
-            hsp_indices.append(i)
-
-        # Calculate k-point distances for band sampling
-        kdistances = []
-        kpoints_abs = kpath["explicit_kpoints_abs"]
-        for i in range(len(hsp_indices) - 1):
-            s = hsp_indices[i]
-            e = hsp_indices[i + 1]
-            if e in discon_hsps:
-                kdistances.append(0)
-            else:
-                kdistances.append(np.linalg.norm(kpoints_abs[e] - kpoints_abs[s]))
-
-        # Determine number of k-points per segment
-        kdistances = np.array(kdistances)
-        min_distance = np.min(kdistances[np.nonzero(kdistances)])
-        band_nks = [int(10 * d / min_distance) for d in kdistances]
-        self.band_nks = np.append(band_nks, 10)
-
-    def write_pwscf_in(self, pwscf_in: str) -> None:
-        """
-        Write Quantum ESPRESSO input file.
-
-        This method writes a complete Quantum ESPRESSO input file with all
-        the necessary namelists and sections for the calculation.
-
-        Parameters
-        ----------
-        pwscf_in : str
-            Path to the output Quantum ESPRESSO input file.
-
-        Examples
-        --------
-        >>> self.write_pwscf_in("scf.in")
-        """
-        with open(pwscf_in, "w") as fp:
-            # Add header comment
-            fp.write(f"! generated by {__file__.split('/')[-1]}\n")
-
-            # Write namelists
-            fp.write(self.control_str + "/\n")
-            fp.write(self.system_str + "/\n")
-            fp.write(self.electrons_str + "/\n")
-
-            # Write structural information
-            fp.write(self.cellparam_str + "\n")
-            fp.write(self.pseudo_str + "\n")
-            fp.write(self.atompos_str + "\n")
-            fp.write(self.kpoints_str + "\n")
-
-    def write_band_in(self, band_in: str) -> None:
-        """
-        Write Quantum ESPRESSO bands input file.
-
-        This method creates a bands input file for post-processing
-        band structure calculations.
-
-        Parameters
-        ----------
-        band_in : str
-            Path to the output bands input file.
-
-        Examples
-        --------
-        >>> self.write_band_in("band/band.in")
-        """
-        with open(band_in, "w") as fp:
-            fp.write(f"! generated by {__file__.split('/')[-1]}\n")
-            fp.write("&bands\n")
-            fp.write(" prefix = 'pwscf'\n")
-            fp.write(" outdir = './work/'\n")
-            fp.write(" filband = 'bands.out'\n")
-            fp.write("/\n")
-
-    def write_pw2wan(self, pw2wan: str) -> None:
-        """
-        Write pw2wannier90 input file.
-
-        This method creates the pw2wannier90 input file that interfaces
-        between Quantum ESPRESSO and Wannier90.
-
-        Parameters
-        ----------
-        pw2wan : str
-            Path to the output pw2wannier90 input file.
-
-        Examples
-        --------
-        >>> self.write_pw2wan("pw2wan.in")
-        """
-        info = self.info_pw2wan
-        with open(pw2wan, "w") as fp:
-            fp.write(f"! generated by {__file__.split('/')[-1]}\n")
-            fp.write("&inputpp\n")
-            fp.write(" outdir = './work'\n")
-            fp.write(" prefix = 'pwscf'\n")
-            fp.write(" seedname = 'pwscf'\n")
-            fp.write(" spin_component = 'none'\n")
-
-            # Add optional wannier_plot_supercell parameter
-            if "wannier_plot_supercell" in info.keys():
-                fp.write(
-                    f" wannier_plot_supercell = {info['wannier_plot_supercell']}\n"
-                )
-
-            # Set output options
-            fp.write(" write_mmn = .true.\n")
-            fp.write(" write_amn = .true.\n")
-            fp.write(f" write_unk = {info['write_unk']}\n")
-            fp.write("/\n")
-
-    def write_wannier(self, wannier_in: str) -> None:
-        """
-        Write Wannier90 input file.
-
-        This method creates a complete Wannier90 input file with all
-        necessary parameters for Wannier function calculation and
-        band structure interpolation.
-
-        Parameters
-        ----------
-        wannier_in : str
-            Path to the output Wannier90 input file.
-
-        Examples
-        --------
-        >>> self.write_wannier("pwscf.win")
-        """
-        with open(wannier_in, "w") as fp:
-            fp.write(f"! generated by {__file__.split('/')[-1]}\n")
-
-            # Determine spin factor for spin-orbit/magnetic calculations
-            so_factor = 1
-            if self.so or self.mag:
-                so_factor = 2
-
-            # Set number of bands and Wannier functions
-            fp.write(f"num_bands = {self.num_wann*3*so_factor}\n")
-            fp.write(f"num_wann  = {self.num_wann*so_factor}\n")
-
-            # Set excluded bands if any
-            if self.nexclude > 0:
-                fp.write(f"exclude_bands = 1-{self.nexclude*so_factor}\n\n")
-
-            # Set disambiguation parameters
-            fp.write("dis_num_iter = 200\n")
-            fp.write("num_iter = 0\n\n")
-            fp.write("dis_froz_max = -200\n")
-            fp.write("dis_froz_min = -200\n\n")
-
-            # Set spinor flag for spin calculations
-            if self.so or self.mag:
-                fp.write("spinors = .true.\n\n")
-
-            # Write projection section
-            fp.write("begin projections\n")
-            fp.write(self.projection_str)
-            fp.write("end projections\n\n")
-
-            # Set output options
-            fp.write("bands_plot = .true.\n")
-            fp.write("write_hr = .true.\n")
-            fp.write("write_tb = .true.\n")
-            fp.write("fermi_surface_plot = .true.\n")
-            fp.write("wannier_plot = .true.\n")
-            fp.write("\n")
-
-            # Write unit cell in Cartesian coordinates
-            fp.write("begin unit_cell_cart\n")
-            fp.write("ang\n")
-            fp.write(
-                f"{self.a1[0]*self.alat:12.7f} {self.a1[1]*self.alat:12.7f} {self.a1[2]*self.alat:12.7f}\n"
-            )
-            fp.write(
-                f"{self.a2[0]*self.alat:12.7f} {self.a2[1]*self.alat:12.7f} {self.a2[2]*self.alat:12.7f}\n"
-            )
-            fp.write(
-                f"{self.a3[0]*self.alat:12.7f} {self.a3[1]*self.alat:12.7f} {self.a3[2]*self.alat:12.7f}\n"
-            )
-            fp.write("end unit_cell_cart\n\n")
-
-            # Write atomic positions
-            fp.write("begin atoms_frac\n")
-            fp.write(self.wan_atompos_str)
-            fp.write("end atoms_frac\n\n")
-
-            # Write k-point mesh
-            fp.write(f"mp_grid: {self.nscfk[0]} {self.nscfk[1]} {self.nscfk[2]}\n\n")
-            fp.write("begin kpoints\n")
-            fp.write(self.wan_kmesh)
-            fp.write("end kpoints\n\n")
-
-            # Write k-point path for band structure
-            fp.write("begin kpoint_path\n")
-            for i in range(len(self.tick_labels) - 1):
-                if self.tick_labels[i] != "" and self.tick_labels[i + 1] != "":
-                    fp.write(
-                        f"{self.tick_labels[i]} {self.tick_locs[i][0]:14.10f} "
-                        f"{self.tick_locs[i][1]:14.10f} {self.tick_locs[i][2]:14.10f}  "
-                        f"{self.tick_labels[i+1]} {self.tick_locs[i+1][0]:14.10f} "
-                        f"{self.tick_locs[i+1][1]:14.10f} {self.tick_locs[i+1][2]:14.10f}\n"
-                    )
-            fp.write("end kpoint_path\n")
-
-    def write_proj(self, proj_in):
-        with open(proj_in, "w") as fp:
-            fp.write("! generated by {}\n".format(__file__.split("/")[-1]))
-            fp.write("&projwfc\n")
-            fp.write(" prefix = 'pwscf'\n")
-            fp.write(" outdir = './work'\n")
-            fp.write(" kresolveddos = .false.\n")
-            fp.write(" degauss = {:0.3f}\n".format(self.degauss))
-            fp.write(" Emax = \n")
-            fp.write(" Emin = \n")
-            fp.write("/\n")
-
-    def write_pp(self, pp_in):
-        so_factor = 1
-        if self.so or self.mag:
-            so_factor = 2
-        with open(pp_in, "w") as fp:
-            fp.write("! generated by {}\n".format(__file__.split("/")[-1]))
-            fp.write("&inputpp\n")
-            fp.write(" prefix = 'pwscf'\n")
-            fp.write(" outdir = './work'\n")
-            fp.write(" filplot = 'wf_pp'\n")
-            fp.write(" plot_num = 7\n")
-            fp.write(" kpoint = 1\n")
-            fp.write(" kband(1) = {}\n".format(self.nexclude * so_factor + 1))
-            fp.write(
-                " kband(2) = {}\n".format(
-                    self.num_wann * so_factor + self.nexclude * so_factor
-                )
-            )
-            fp.write(" lsign = .TRUE.\n")
-            fp.write("/\n")
-            fp.write("&plot\n")
-            fp.write(" iflag = 3\n")
-            fp.write(" output_format = 5\n")
-            fp.write(" fileout = '.xsf'\n")
-            fp.write("/\n")
-
-
-class pseudo_list:
-    """
-    Handle pseudopotential information and selection.
-
-    This class manages pseudopotential files and their associated parameters
-    for Quantum ESPRESSO calculations. It reads pseudopotential information
-    from a CSV file and provides methods to retrieve pseudopotential data.
-
-    Parameters
-    ----------
-    pseudo_dir : str
-        Directory containing pseudopotential files.
-    pp_list_path : str
-        Path to CSV file containing pseudopotential information.
-
-    Attributes
-    ----------
-    pseudo_dir : str
-        Directory containing pseudopotential files.
-    pp_info : dict
-        Dictionary mapping atom types to pseudopotential information.
-    """
-
-    def __init__(self, pseudo_dir: str, pp_list_path: str) -> None:
-        """
-        Initialize pseudopotential list handler.
-
-        Parameters
-        ----------
-        pseudo_dir : str
-            Directory containing pseudopotential files.
-        pp_list_path : str
-            Path to CSV file containing pseudopotential information.
-        """
-        self.pseudo_dir = pseudo_dir
-        self.pp_info = self.read_pp_info(pp_list_path)
-
-    def pseudo(self, atm: str) -> tuple:
-        """
-        Get pseudopotential information for an atom type.
-
-        Parameters
-        ----------
-        atm : str
-            Atom type symbol.
-
-        Returns
-        -------
-        tuple
-            Pseudopotential information tuple containing:
-            - pp_file_name: Name of pseudopotential file
-            - nexclude: Number of excluded bands
-            - orbital: Orbital types for projections
-            - num_wann: Number of Wannier functions
-            - ecutwfc: Wavefunction cutoff energy
-            - ecutrho: Charge density cutoff energy
-        """
-        return self.pp_info[atm]
-
-    def read_pp_info(self, pp_list_path: str) -> dict:
-        """
-        Read pseudopotential information from CSV file.
-
-        Parameters
-        ----------
-        pp_list_path : str
-            Path to CSV file containing pseudopotential information.
-
-        Returns
-        -------
-        dict
-            Dictionary mapping atom types to pseudopotential information.
-        """
-        csv_input = pd.read_csv(pp_list_path, sep=",").fillna("NaN")
-        pp_dict = {}
-
-        for value in csv_input.values:
-            if value[1] == "NaN":  # No pseudopotential file
-                pp_dict[value[0]] = None
-            else:
-                # Handle missing values
-                if value[2] == "NaN":  # nexclude
-                    value[2] = 0
-                if value[3] == "NaN":  # orbital
-                    value[3] = ""
-                if value[4] == "NaN":  # ecutwfc
-                    value[4] = 0
-                if value[5] == "NaN":  # ecutrho
-                    value[5] = 0
-
-                # Create pseudopotential filename
-                pp_file_name = f"{value[1]}.UPF"
-
-                # Calculate number of Wannier functions from orbital types
-                num_wann = 0
-                for s in value[3]:
-                    if s == "s":
-                        num_wann += 1
-                    elif s == "p":
-                        num_wann += 3
-                    elif s == "d":
-                        num_wann += 5
-                    elif s == "f":
-                        num_wann += 7
-
-                # Store pseudopotential information
-                pp_dict[value[0]] = (
-                    pp_file_name,
-                    int(value[2]),
-                    value[3],
-                    num_wann,
-                    float(value[4]),
-                    float(value[5]),
-                )
-        return pp_dict
-
-
-def main():
-    """
-    Main function for cif2qewan command-line interface.
-
-    This function provides the entry point for the cif2qewan command
-    and handles the complete workflow from CIF file to Quantum ESPRESSO
-    and Wannier90 input files.
-    """
-    # Parse command-line arguments
-    args = docopt(__doc__)
-
-    # Extract input parameters
-    cif_file = args["<cif_file>"]
-    toml_file = args["<toml_file>"]
-
-    # Initialize the QE Wannier input generator
-    qe_wan = qe_wannier_in(cif_file, toml_file, args["--so"], args["--mag"])
-
-    # Generate SCF input file
-    qe_wan.write_pwscf_in("scf.in")
-
-    # Convert to NSCF for Wannier90
-    qe_wan.convert2nscf()
-    qe_wan.write_pwscf_in("nscf.in")
-
-    # Calculate band structure k-path
-    qe_wan.calc_bands_seekpath()
-
-    # Generate interface files
-    qe_wan.write_pw2wan("pw2wan.in")
-    qe_wan.write_wannier("pwscf.win")
-
-    # Create check_wannier directory for convergence testing
-    if not os.path.exists("check_wannier"):
-        os.mkdir("check_wannier")
-    qe_wan.shift_k_nscf()
-    qe_wan.write_pwscf_in("check_wannier/nscf.in")
-
-    # Create band directory for band structure calculations
-    if not os.path.exists("band"):
-        os.mkdir("band")
-    qe_wan.convert2band()
-    qe_wan.write_pwscf_in("band/nscf.in")
-    qe_wan.write_band_in("band/band.in")
-    qe_wan.write_proj("band/proj.in")
-    qe_wan.write_pp("band/pp.in")
-
-
-if __name__ == "__main__":
-    """
-    Main execution block for CIF to Quantum ESPRESSO Wannier90 converter.
-
-    This script generates all necessary input files for a complete
-    Quantum ESPRESSO + Wannier90 calculation workflow from a CIF file.
-    """
-    main()
