+++ cif2qewan/__init__.py
@@ -1,31 +1,15 @@
-# -*- coding: utf-8 -*-
 """
-cif2qewan: A comprehensive Python toolkit for Quantum ESPRESSO and Wannier90.
+cif2qewan: Quantum ESPRESSO and Wannier90 input generation from CIF files.
 
-This package provides tools for generating Quantum ESPRESSO and Wannier90 input files
-from CIF (Crystallographic Information File) structures, with automated workflow
-management and band structure analysis capabilities.
-
-Main modules:
-- cif2qewan: Main CIF to QE/Wannier90 converter
-- band_comp: Band structure comparison tool
-- wannier_conv: Wannier90 convergence checker
-- get_cif: Materials Project CIF downloader
-
-Author: Generated by cif2qewan
-Version: 1.0.0
-License: MIT
+Modules
+-------
+cif2qewan
+    Generate the Quantum ESPRESSO and Wannier90 input files from a CIF file.
+band_comp
+    Compare DFT and Wannier90 band structures.
+wannier_conv
+    Check the accuracy of the Wannier90 interpolation.
 """
 
-__version__ = "1.0.0"
-__author__ = "cif2qewan developers"
-__email__ = "cif2qewan@example.com"
-__license__ = "MIT"
-
-# Define what gets imported with "from cif2qewan import *"
-__all__ = [
-    "__version__",
-    "__author__",
-    "__email__",
-    "__license__",
-]
+__version__ = "0.1.0"
+__all__ = ["__version__"]
