--- nc-sr-05_pbe_standard_upf.csv
@@ -1,119 +1,0 @@
-atom,pp_file_name,nexclude,orbitals,ecutwfc,ecutrho
-H,H_sr,0,s,84.0,336.0
-He,He_sr,1,,98.0,392.0
-Li,Li_sr,1,s,82.0,328.0
-Be,Be_sr,1,sp,100.0,400.0
-B,B_sr,0,sp,88.0,352.0
-C,C_sr,0,sp,90.0,360.0
-N,N_sr,1,p,96.0,384.0
-O,O_sr,1,p,96.0,384.0
-F,F_sr,1,p,96.0,384.0
-Ne,Ne_sr,1,p,80.0,320.0
-Na,Na_sr,4,sp,96.0,384.0
-Mg,Mg_sr,4,sp,96.0,384.0
-Al,Al_sr,0,sp,52.0,208.0
-Si,Si_sr,0,sp,48.0,192.0
-P,P_sr,1,p,56.0,224.0
-S,S_sr,1,p,64.0,256.0
-Cl,Cl_sr,1,p,66.0,264.0
-Ar,Ar_sr,1,p,74.0,296.0
-K,K_sr,4,s,86.0,344.0
-Ca,Ca_sr,4,s,76.0,304.0
-Sc,Sc_sr,4,sd,90.0,360.0
-Ti,Ti_sr,4,sdp,92.0,368.0
-V,V_sr,4,ds,96.0,384.0
-Cr,Cr_sr,4,sd,110.0,440.0
-Mn,Mn_sr,4,ds,108.0,432.0
-Fe,Fe_sr,4,dsp,106.0,424.0
-Co,Co_sr,4,dsp,108.0,432.0
-Ni,Ni_sr,4,dsp,110.0,440.0
-Cu,Cu_sr,4,dsp,104.0,416.0
-Zn,Zn_sr,4,dsp,96.0,384.0
-Ga,Ga_sr,5,sp,92.0,368.0
-Ge,Ge_sr,5,sp,90.0,360.0
-As,As_sr,6,p,96.0,384.0
-Se,Se_sr,6,p,98.0,392.0
-Br,Br_sr,1,p,58.0,232.0
-Kr,Kr_sr,1,p,68.0,272.0
-Rb,Rb_sr,4,sp,58.0,232.0
-Sr,Sr_sr,4,sp,80.0,320.0
-Y,Y_sr,4,sdp,84.0,336.0
-Zr,Zr_sr,4,sdp,98.0,392.0
-Nb,Nb_sr,4,sd,98.0,392.0
-Mo,Mo_sr,4,sdp,92.0,368.0
-Tc,Tc_sr,4,dsp,96.0,384.0
-Ru,Ru_sr,4,dsp,100.0,400.0
-Rh,Rh_sr,4,dsp,100.0,400.0
-Pd,Pd_sr,4,dsp,98.0,392.0
-Ag,Ag_sr,4,dsp,94.0,376.0
-Cd,Cd_sr,9,sp,114.0,456.0
-In,In_sr,5,sp,82.0,328.0
-Sn,Sn_sr,5,sp,84.0,336.0
-Sb,Sb_sr,5,sp,88.0,352.0
-Te,Te_sr,6,p,92.0,368.0
-I,I_sr,1,p,82.0,328.0
-Xe,Xe_sr,1,p,84.0,336.0
-Cs,Cs_sr,1,ps,58.0,232.0
-Ba,Ba_sr,4,sp,56.0,224.0
-La,La_sr,4,fsdp,130.0,520.0
-Ce,Ce_sr,,,,
-Pr,Pr_sr,,,,
-Nd,Nd_sr,,,,
-Pm,Pm_sr,,,,
-Sm,Sm_sr,,,,
-Eu,Eu_sr,,,,
-Gd,Gd_sr,,,,
-Tb,Tb_sr,,,,
-Dy,Dy_sr,,,,
-Ho,Ho_sr,,,,
-Er,Er_sr,,,,
-Tm,Tm_sr,,,,
-Yb,Yb_sr,,,,
-Lu,Lu_sr,11,sdp,116.0,464.0
-Hf,Hf_sr,4,sdp,70.0,280.0
-Ta,Ta_sr,4,sdp,70.0,280.0
-W,W_sr,4,sdp,82.0,328.0
-Re,Re_sr,4,sdp,84.0,336.0
-Os,Os_sr,4,dsp,86.0,344.0
-Ir,Ir_sr,4,dsp,80.0,320.0
-Pt,Pt_sr,4,dsp,100.0,400.0
-Au,Au_sr,4,dsp,88.0,352.0
-Hg,Hg_sr,4,dsp,78.0,312.0
-Tl,Tl_sr,5,sp,74.0,296.0
-Pb,Pb_sr,5,sp,68.0,272.0
-Bi,Bi_sr,6,p,74.0,296.0
-Po,Po_sr,6,p,76.0,304.0
-At,At_sr,,,,
-Rn,Rn_sr,6,p,84.0,336.0
-Fr,Fr_sr,,,,
-Ra,Ra_sr,,,,
-Ac,Ac_sr,,,,
-Th,Th_sr,,,,
-Pa,Pa_sr,,,,
-U,U_sr,,,,
-Np,Np_sr,,,,
-Pu,Pu_sr,,,,
-Am,Am_sr,,,,
-Cm,Cm_sr,,,,
-Bk,Bk_sr,,,,
-Cf,Cf_sr,,,,
-Es,Es_sr,,,,
-Fm,Fm_sr,,,,
-Md,Md_sr,,,,
-No,No_sr,,,,
-Lr,Lr_sr,,,,
-Rf,Rf_sr,,,,
-Db,Db_sr,,,,
-Sg,Sg_sr,,,,
-Bh,Bh_sr,,,,
-Hs,Hs_sr,,,,
-Mt,Mt_sr,,,,
-Ds,Ds_sr,,,,
-Rg,Rg_sr,,,,
-Cn,Cn_sr,,,,
-Nh,Nh_sr,,,,
-Fl,Fl_sr,,,,
-Mc,Mc_sr,,,,
-Lv,Lv_sr,,,,
-Ts,Ts_sr,,,,
-Og,Og_sr,,,,
