--- pp_psl_rrkj.csv
@@ -1,110 +1,0 @@
-atom,pp_file_name,nexclude,orbitals,ecutwfc,ecutrho
-H,H.pbe-rrkjus_psl.0.1,0,s,46.0,221.0
-He,He.pbe-rrkjus_psl.0.2.3,1,,41.0,165.0
-Li,Li.pbe-s-rrkjus_psl.0.2.1,1,s,78.0,355.0
-Be,Be.pbe-n-rrkjus_psl.0.2,0,sp,30.0,197.0
-B,B.pbe-n-rrkjus_psl.0.1,0,sp,48.0,216.0
-C,C.pbe-n-rrkjus_psl.0.1,0,sp,37.0,318.0
-N,N.pbe-n-rrkjus_psl.0.1,1,p,39.0,264.0
-O,O.pbe-n-rrkjus_psl.0.1,1,p,47.0,323.0
-F,F.pbe-n-rrkjus_psl.0.1,1,p,48.0,316.0
-Ne,Ne.pbe-n-rrkjus_psl.1.0.0,1,p,54.0,264.0
-Na,Na.pbe-spn-rrkjus_psl.0.2,4,s,41.0,163.0
-Mg,Mg.pbe-n-rrkjus_psl.0.3.0,0,s,13.0,87.0
-Al,Al.pbe-n-rrkjus_psl.0.1,0,sp,36.0,145.0
-Si,Si.pbe-n-rrkjus_psl.0.1,0,sp,38.0,151.0
-P,P.pbe-n-rrkjus_psl.0.1,1,p,17.0,79.0
-S,S.pbe-n-rrkjus_psl.0.1,1,p,17.0,77.0
-Cl,Cl.pbe-n-rrkjus_psl.0.3.0,1,p,36.0,146.0
-Ar,Ar.pbe-n-rrkjus_psl.0.3.0,1,p,26.0,225.0
-K,K.pbe-spn-rrkjus_psl.0.2.3,4,s,34.0,366.0
-Ca,Ca.pbe-spn-rrkjus_psl.0.2.3,4,s,36.0,170.0
-Sc,Sc.pbe-spn-rrkjus_psl.0.2.3,4,sd,46.0,477.0
-Ti,Ti.pbe-spn-rrkjus_psl.0.3.1,4,spd,50.0,738.0
-V,V.pbe-spn-rrkjus_psl.0.2.3,4,sd,47.0,731.0
-Cr,Cr.pbe-spn-rrkjus_psl.0.2.3,4,spd,50.0,239.0
-Mn,Mn.pbe-spn-rrkjus_psl.0.3.1,4,spd,46.0,244.0
-Fe,Fe.pbe-spn-rrkjus_psl.0.2.1,4,spd,64.0,782.0
-Co,Co.pbe-n-rrkjus_psl.0.2.3,0,spd,36.0,180.0
-Ni,Ni.pbe-n-rrkjus_psl.0.1,0,spd,41.0,236.0
-Cu,Cu.pbe-dn-rrkjus_psl.0.2,0,spd,33.0,150.0
-Zn,Zn.pbe-dn-rrkjus_psl.0.2.2,0,spd,39.0,232.0
-Ga,Ga.pbe-dn-rrkjus_psl.0.2,5,sp,33.0,180.0
-Ge,Ge.pbe-dn-rrkjus_psl.0.3.1,5,sp,36.0,240.0
-As,As.pbe-n-rrkjus_psl.0.2,1,p,20.0,103.0
-Se,Se.pbe-n-rrkjus_psl.0.2,1,p,21.0,95.0
-Br,Br.pbe-n-rrkjus_psl.0.2,1,p,25.0,108.0
-Kr,Kr.pbe-n-rrkjus_psl.0.2.3,1,p,28.0,220.0
-Rb,Rb.pbe-spn-rrkjus_psl.0.2.3,4,s,28.0,217.0
-Sr,Sr.pbe-spn-rrkjus_psl.0.2.3,4,s,31.0,380.0
-Y,Y.pbe-spn-rrkjus_psl.0.2.3,4,spd,36.0,252.0
-Zr,Zr.pbe-spn-rrkjus_psl.0.2.3,4,spd,41.0,326.0
-Nb,Nb.pbe-spn-rrkjus_psl.0.3.0,4,sd,42.0,364.0
-Mo,Mo.pbe-spn-rrkjus_psl.0.3.0,4,spd,48.0,449.0
-Tc,Tc.pbe-spn-rrkjus_psl.0.3.0,4,spd,62.0,832.0
-Ru,Ru.pbe-spn-rrkjus_psl.0.3.0,4,spd,62.0,798.0
-Rh,Rh.pbe-spn-rrkjus_psl.0.3.0,4,spd,55.0,439.0
-Pd,Pd.pbe-n-rrkjus_psl.0.3.0,0,sd,32.0,138.0
-Ag,Ag.pbe-dn-rrkjus_psl.0.1,0,sd,32.0,165.0
-Cd,Cd.pbe-dn-rrkjus_psl.0.3.1,5,sp,37.0,179.0
-In,In.pbe-dn-rrkjus_psl.0.2.2,5,sp,48.0,190.0
-Sn,Sn.pbe-dn-rrkjus_psl.0.2,5,sp,25.0,126.0
-Sb,Sb.pbe-n-rrkjus_psl.0.3.1,0,sp,33.0,132.0
-Te,Te.pbe-dn-rrkjus_psl.0.3.1,6,p,43.0,211.0
-I,I.pbe-n-rrkjus_psl.0.2,1,p,28.0,113.0
-Xe,Xe.pbe-n-rrkjus_psl.0.3.1,6,p,42.0,210.0
-Cs,Cs.pbe-spn-rrkjus_psl.0.2.3,1,sp,35.0,150.0
-Ba,Ba.pbe-spn-rrkjus_psl.0.2.3,4,s,28.0,166.0
-La,La.pbe-spfn-rrkjus_psl.1.0.0,4,spdf,75.0,537.0
-Ce,Ce.pbe-spdn-rrkjus_psl.1.0.0,4,spd,37.0,179.0
-Pr,Pr.pbe-spdn-rrkjus_psl.1.0.0,4,spd,41.0,309.0
-Nd,Nd.pbe-spdn-rrkjus_psl.1.0.0,4,spd,38.0,202.0
-Pm,Pm.pbe-spdn-rrkjus_psl.1.0.0,4,spd,34.0,198.0
-Sm,Sm.pbe-spdn-rrkjus_psl.1.0.0,4,spd,34.0,195.0
-Eu,Eu.pbe-spdn-rrkjus_psl.1.0.0,4,spd,35.0,259.0
-Gd,Gd.pbe-spdn-rrkjus_psl.1.0.0,4,spd,35.0,227.0
-Tb,Tb.pbe-spdn-rrkjus_psl.1.0.0,4,spd,35.0,229.0
-Dy,Dy.pbe-spdn-rrkjus_psl.1.0.0,4,spd,36.0,265.0
-Ho,Ho.pbe-spdn-rrkjus_psl.1.0.0,4,spd,36.0,268.0
-Er,Er.pbe-spdn-rrkjus_psl.1.0.0,4,spd,37.0,265.0
-Tm,Tm.pbe-spdn-rrkjus_psl.1.0.0,4,spd,38.0,316.0
-Yb,Yb.pbe-spdn-rrkjus_psl.1.0.0,4,spd,39.0,278.0
-Lu,Lu.pbe-spdn-rrkjus_psl.1.0.0,4,spd,42.0,333.0
-Hf,Hf.pbe-spn-rrkjus_psl.0.2,4,spd,43.0,311.0
-Ta,Ta.pbe-spn-rrkjus_psl.0.2,4,spd,47.0,387.0
-W,W.pbe-spn-rrkjus_psl.0.2.3,4,spd,49.0,494.0
-Re,Re.pbe-spn-rrkjus_psl.0.3.1,4,spd,54.0,374.0
-Os,Os.pbe-n-rrkjus_psl.0.2.3,0,spd,26.0,261.0
-Ir,Ir.pbe-n-rrkjus_psl.0.2.3,0,spd,27.0,361.0
-Pt,Pt.pbe-n-rrkjus_psl.0.1,0,spd,39.0,401.0
-Au,Au.pbe-dn-rrkjus_psl.0.3.0,0,spd,26.0,357.0
-Hg,Hg.pbe-dn-rrkjus_psl.0.2.2,0,spd,29.0,119.0
-Tl,Tl.pbe-dn-rrkjus_psl.0.2.3,5,sp,35.0,150.0
-Pb,Pb.pbe-dn-rrkjus_psl.0.2.2,5,sp,40.0,158.0
-Bi,Bi.pbe-dn-rrkjus_psl.0.2.2,6,p,31.0,443.0
-Po,Po.pbe-dn-rrkjus_psl.0.3.1,6,p,53.0,447.0
-At,At.pbe-dn-rrkjus_psl.0.3.1,6,p,63.0,463.0
-Rn,Rn.pbe-dn-rrkjus_psl.0.3.1,6,p,57.0,465.0
-Fr,,,,,
-Ra,,,,,
-Ac,,,,,
-Th,,,,,
-Pa,,,,,
-U,,,,,
-Np,,,,,
-Pu,,,,,
-Am,,,,,
-Cm,,,,,
-Bk,,,,,
-Cf,,,,,
-Es,,,,,
-Fm,,,,,
-Md,,,,,
-No,,,,,
-Lr,,,,,
-Rf,,,,,
-Db,,,,,
-Sg,,,,,
-Bh,,,,,
-Hs,,,,,
-Mt,,,,,
