--- wannier_conv.py
@@ -1,441 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Wannier90 convergence checker.
-
-This module provides functionality to check the accuracy of Wannier90
-interpolation by comparing the interpolated band structure with the
-original DFT band structure from Quantum ESPRESSO calculations.
-
-Author: Generated by cif2qewan
-"""
-
-# Standard library imports
-import argparse
-import itertools
-import os
-from typing import Tuple
-
-# Third-party imports
-import numpy as np
-
-
-class Hamiltonian:
-    """
-    Wannier90 Hamiltonian data handler.
-
-    This class contains the information of Hamiltonian generated by wannier90 code.
-    All units are eV and A. It reads the hr.dat file and provides methods for
-    diagonalizing the Hamiltonian at arbitrary k-points.
-
-    Attributes
-    ----------
-    num_wann : int
-        Number of Wannier orbitals.
-    nrpts : int
-        Number of R vectors in the Wigner-Seitz grid.
-    ndegen : np.ndarray
-        Degeneracies of Wigner-Seitz grid points (shape: [nrpts]).
-    irvec : np.ndarray
-        Lattice vectors for unit cells (shape: [3, nrpts]).
-    ir0 : int
-        Index of H(R=0). Returns -1 if not found.
-    ham_r : np.ndarray
-        Wannier-based Hamiltonian (shape: [num_wann, num_wann, nrpts]).
-    reorder : bool
-        Whether to reorder indices of Wannier orbitals.
-
-    Examples
-    --------
-    >>> h = Hamiltonian("pwscf_hr.dat")
-    >>> ek, v = h.diagonalize([0.0, 0.0, 0.0])
-    """
-
-    def __init__(self, file_hr: str, reorder: bool = False) -> None:
-        """
-        Initialize Hamiltonian from hr.dat file.
-
-        Parameters
-        ----------
-        file_hr : str
-            Path to the hr.dat file generated by Wannier90.
-        reorder : bool, optional
-            Whether to reorder indices of Wannier orbitals in dat file.
-            Used for compatibility with VASP and older Wannier90 versions.
-            Default is False.
-
-        Examples
-        --------
-        >>> h = Hamiltonian("pwscf_hr.dat")
-        >>> h_reordered = Hamiltonian("pwscf_hr.dat", reorder=True)
-        """
-        self.reorder = reorder
-        self._read_hr(file_hr)
-
-    def _read_hr(self, file_hr: str) -> None:
-        """
-        Read Hamiltonian data from hr.dat file.
-
-        This method parses the Wannier90 hr.dat file format and loads
-        the Hamiltonian matrix elements and lattice vectors.
-
-        Parameters
-        ----------
-        file_hr : str
-            Path to the hr.dat file.
-
-        Raises
-        ------
-        Exception
-            If the file cannot be read or parsed.
-        """
-        try:
-            # Check if file exists
-            if os.path.exists(file_hr):
-                fp = open(file_hr, "r")
-            else:
-                raise FileNotFoundError(f"File {file_hr} not found")
-
-            # Skip empty line
-            fp.readline()
-
-            # Read number of Wannier orbitals and R points
-            self.num_wann = int(fp.readline())
-            self.nrpts = int(fp.readline())
-
-            # Read degeneracies (15 per line)
-            ndegen = []
-            for i in range(int(self.nrpts / 15) + 1):
-                ndegen += [int(x) for x in fp.readline().split()]
-                if len(ndegen) >= self.nrpts:
-                    break
-            self.ndegen = np.array(ndegen)
-
-            # Initialize Hamiltonian and lattice vectors
-            self.ham_r = np.zeros(
-                (self.num_wann, self.num_wann, self.nrpts), dtype=np.complex128
-            )
-            self.irvec = np.zeros((3, self.nrpts), dtype=np.int64)
-
-            # Find R=0 index
-            self.ir0 = -1
-
-            # Read Hamiltonian matrix elements
-            for i, m, n in itertools.product(
-                range(self.nrpts), range(self.num_wann), range(self.num_wann)
-            ):
-                (irx, iry, irz, _, _, tr, ti) = fp.readline().split()
-
-                # Store lattice vectors (only once per R point)
-                if m == 0 and n == 0:
-                    self.irvec[0:3, i] = np.array([int(x) for x in [irx, iry, irz]])
-                    if np.all(self.irvec[0:3, i] == 0):
-                        self.ir0 = i
-
-                # Store Hamiltonian matrix elements
-                if self.reorder:
-                    # VASP, wannier ver1. etc - reorder indices
-                    mr = m // 2 + (self.num_wann // 2) * (m % 2)
-                    nr = n // 2 + (self.num_wann // 2) * (n % 2)
-                    self.ham_r[nr, mr, i] = float(tr) + float(ti) * 1j
-                else:
-                    # Normal order
-                    self.ham_r[n, m, i] = float(tr) + float(ti) * 1j
-
-            fp.close()
-
-        except Exception as e:
-            print(f"Failed to read: {file_hr}")
-            print(f"Error type: {type(e)}")
-            print(f"Error args: {e.args}")
-            print(str(e))
-
-    def diagonalize(self, k: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
-        """
-        Diagonalize the Hamiltonian at a given k-point.
-
-        This method constructs the k-space Hamiltonian H(k) from the
-        real-space Hamiltonian H(R) using the Fourier transform:
-        H(k) = Σ_R H(R) * exp(i*k·R) / deg(R)
-
-        Parameters
-        ----------
-        k : np.ndarray
-            K-point in crystal coordinates (shape: [3]).
-
-        Returns
-        -------
-        tuple[np.ndarray, np.ndarray]
-            A tuple containing:
-            - e: Eigenvalues (energies) in eV
-            - v: Eigenvectors (wavefunctions)
-
-        Examples
-        --------
-        >>> h = Hamiltonian("pwscf_hr.dat")
-        >>> ek, v = h.diagonalize([0.0, 0.0, 0.0])
-        >>> print(f"Energies: {ek}")
-        """
-        # Calculate k·R for all R vectors
-        kr = np.dot(k, self.irvec)
-
-        # Fourier transform factor: exp(i*k·R) / deg(R)
-        pi = np.pi
-        factor = np.exp(2 * pi * 1j * kr) / self.ndegen
-
-        # Construct k-space Hamiltonian
-        ham = np.dot(self.ham_r, factor)
-
-        # Diagonalize to get eigenvalues and eigenvectors
-        (e, v) = np.linalg.eigh(ham)
-
-        return (e, v)
-
-
-class Nscfout:
-    """
-    Quantum ESPRESSO NSCF output parser.
-
-    This class parses the output from Quantum ESPRESSO NSCF calculations
-    to extract k-points, energies, and other relevant information.
-    The NSCF calculation must be done with "verbosity = 'high'" for
-    proper parsing of the output.
-
-    Attributes
-    ----------
-    ef : float
-        Fermi energy in eV.
-    nbnd : int
-        Number of Kohn-Sham bands.
-    nk : int
-        Number of k-points.
-    kp_cart : np.ndarray
-        K-points in Cartesian coordinates (shape: [nk, 3]).
-    kp_cryst : np.ndarray
-        K-points in crystal coordinates (shape: [nk, 3]).
-    wk : np.ndarray
-        K-point weights (shape: [nk]).
-    energy : np.ndarray
-        Band energies (shape: [nk, nbnd]).
-
-    Examples
-    --------
-    >>> nscf_data = Nscfout("nscf.out")
-    >>> print(f"Fermi energy: {nscf_data.ef} eV")
-    """
-
-    def __init__(self, nscf_out: str) -> None:
-        """
-        Parse NSCF output file.
-
-        Parameters
-        ----------
-        nscf_out : str
-            Path to the NSCF output file.
-        """
-        with open(nscf_out, "r") as fp:
-            lines = fp.readlines()
-
-        # Parse header information
-        for i, line in enumerate(lines):
-            if "the Fermi energy" in line:
-                self.ef = float(line[26:35])
-            if "number of Kohn-Sham" in line:
-                self.nbnd = int(line[35:])
-            if "number of k points=" in line:
-                self.nk = int(line[25:31])
-                # Initialize k-point arrays
-                self.kp_cart = np.zeros([self.nk, 3])
-                self.kp_cryst = np.zeros([self.nk, 3])
-                self.wk = np.zeros([self.nk])
-
-                # Parse k-points
-                for j in range(self.nk):
-                    self.kp_cart[j] = np.array(
-                        [float(x) for x in lines[i + j + 2][20:56].split()]
-                    )
-                    self.wk[j] = float(lines[i + j + 2][65:])
-                    self.kp_cryst[j] = np.array(
-                        [float(x) for x in lines[i + j + 4 + self.nk][20:56].split()]
-                    )
-
-        # Parse band energies
-        self.energy = np.zeros([self.nk, self.nbnd])
-        nline, nlinemod = divmod(self.nbnd, 8)
-        if nlinemod > 0:
-            nline += 1
-
-        for j in range(self.nk):
-            # Find k-point in output
-            kp_str = "k =%7.4f%7.4f%7.4f" % tuple(self.kp_cart[j])
-            for i, line in enumerate(lines):
-                if kp_str in line:
-                    # Parse energy values (8 per line)
-                    tmp_list = "".join(lines[i + 2 : i + 2 + nline]).split()
-                    ene_list = []
-
-                    # Handle long energy values that may be concatenated
-                    for l_tmp in tmp_list:
-                        len_max = 8
-                        if len(l_tmp) > len_max:
-                            # Split concatenated values
-                            for s in range(len(l_tmp) // (len_max + 1)):
-                                ene_list.append(
-                                    l_tmp[s * (len_max + 1) : (s + 1) * (len_max + 1)]
-                                )
-                        else:
-                            ene_list.append(l_tmp)
-
-                    # Store energies for this k-point
-                    self.energy[j, :] = [float(x) for x in ene_list]
-
-
-def get_nexclude(pwscf_win: str) -> int:
-    """
-    Extract number of excluded bands from Wannier90 input file.
-
-    Parameters
-    ----------
-    pwscf_win : str
-        Path to the Wannier90 input file (pwscf.win).
-
-    Returns
-    -------
-    int
-        Number of excluded bands.
-
-    Examples
-    --------
-    >>> nexclude = get_nexclude("pwscf.win")
-    >>> print(f"Excluded bands: {nexclude}")
-    """
-    nexclude = 0
-    with open(pwscf_win, "r") as fp:
-        for line in fp.readlines():
-            if "exclude" in line:
-                # Parse exclude_bands = 1-N format
-                nexclude = int(line.split("-")[1])
-                break
-    return nexclude
-
-
-def main():
-    """
-    Main function for wannier_conv command-line interface.
-
-    This function provides the entry point for the wannier_conv command
-    and handles the convergence checking workflow.
-    """
-    # Set up command-line argument parser
-    parser = argparse.ArgumentParser(
-        description="Check accuracy of Wannier90 interpolation",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  python wannier_conv.py -e 5.0 -o ./output
-  python wannier_conv.py -e 0.0 -i ./check_wannier/nscf.out -o ./results
-        """,
-    )
-    parser.add_argument(
-        "-e",
-        type=float,
-        default=0.0,
-        dest="emax",
-        help="Maximum energy above Fermi level for comparison (eV)",
-    )
-    parser.add_argument(
-        "-o",
-        dest="odir",
-        required=True,
-        help="Output directory for convergence results",
-    )
-    parser.add_argument(
-        "-i",
-        default="./check_wannier/nscf.out",
-        dest="nscf_file",
-        help="Path to NSCF output file",
-    )
-
-    # Parse command-line arguments
-    args = parser.parse_args()
-
-    # Set energy window for comparison
-    emin = -200.0  # Minimum energy (well below Fermi level)
-    emax = args.emax  # Maximum energy above Fermi level
-    output_dir = args.odir
-    nscf_file = args.nscf_file
-
-    # Load data from files
-    print("Loading NSCF data...")
-    nscf_data = Nscfout(nscf_file)
-
-    print("Loading Wannier90 Hamiltonian...")
-    nexclude = get_nexclude("./pwscf.win")
-    h = Hamiltonian("./pwscf_hr.dat")
-
-    print(f"Comparing energies in window [{emin:.2f}:{emax:.2f}] eV...")
-
-    # Calculate energy differences between Wannier90 and DFT
-    delta_sum = 0.0
-    delta_max = 0.0
-    nek = 0  # Number of k-points and bands compared
-
-    for i in range(nscf_data.nk):
-        # Get Wannier90 energies at this k-point
-        (ek, _) = h.diagonalize(nscf_data.kp_cryst[i])
-
-        # Determine energy range for comparison
-        nek_low = np.sum(ek - nscf_data.ef < emin)
-        nek_max = np.sum(ek - nscf_data.ef < emax)
-
-        # Ensure we don't exceed available bands
-        if nexclude + nek_max > nscf_data.nbnd:
-            nek_max = nscf_data.nbnd - nexclude
-
-        # Skip if no bands in this energy range
-        if nek_max == nek_low:
-            continue
-
-        # Count bands in comparison
-        nek += nek_max - nek_low
-
-        # Calculate energy differences (squared)
-        ediff = (
-            ek[nek_low:nek_max]
-            - nscf_data.energy[i, nexclude + nek_low : nexclude + nek_max]
-        ) ** 2
-        delta_sum += np.sum(ediff)
-        delta_max = max([delta_max, np.max(ediff)])
-
-    # Write convergence results
-    output_file = os.path.join(output_dir, f"CONV_{emax}")
-    with open(output_file, "w") as fp:
-        # Keep the CONV format unchanged: existing scripts parse these lines.
-        fp.write(f"# energy window [{emin:>5.2f}:{emax:>5.2f}]\n")
-
-        if nek > 0:
-            avg_diff = np.sqrt(delta_sum/nek)
-            fp.write(f"average diff = {avg_diff:>15.8f}\n")
-        else:
-            fp.write("average diff = NaN\n")
-
-        max_diff = np.sqrt(delta_max)
-        fp.write(f"max diff     = {max_diff:>15.8f}\n")
-    
-    print(f"Convergence results written to: {output_file}")
-    if nek > 0:
-        print(f"Average difference: {np.sqrt(delta_sum/nek):.6f} eV")
-        print(f"Maximum difference: {np.sqrt(delta_max):.6f} eV")
-    else:
-        print("No bands found in the specified energy window")
-
-
-if __name__ == "__main__":
-    """
-    Main execution block for Wannier90 convergence checker.
-
-    This script compares the interpolated band structure from Wannier90
-    with the original DFT band structure to check the accuracy of the
-    Wannier interpolation.
-    """
-    main()
